fixed program status word
